adapt conftest bq to run tests locally

+++ conftest.py
@@ -3,6 +3,7 @@ import warnings
 import pytest
 from cuallee import Check, CheckLevel
 from pyspark.sql import SparkSession
+from pathlib import Path
 import logging
 import duckdb
 
@@ -84,14 +85,18 @@ def db() -> duckdb.DuckDBPyConnection:
 
 @pytest.fixture(scope="session")
 def bq_client():
-    from google.oauth2 import service_account
-    import os
-    import json
-    with open("key.json", "w") as writer:
-        json.dump(json.loads(os.getenv("GOOGLE_KEY")), writer)
 
-    credentials = service_account.Credentials.from_service_account_file("key.json")
-    
+    if Path('temp/key.json').exists()==True:
+        credentials = os.getenv('GOOGLE_APPLICATION_CREDENTIALS')
+    else:
+        from google.oauth2 import service_account
+        import os
+        import json
+        with open("key.json", "w") as writer:
+            json.dump(json.loads(os.getenv("GOOGLE_KEY")), writer)
+
+        credentials = service_account.Credentials.from_service_account_file("key.json")
+        
     try:
         client = bigquery.Client(project="cuallee-bigquery-386709", credentials=credentials)
         return client
